Make data type size table available as property
The type size table can simply be indexed by a
datatype variable to get the corresponding size
in bytes. Unfortunately this means we have to
open and close the binary trace twice, since
reading the sizes jumps to the end of the file.

Also add an explicit open() method, as well as
some error checking: reporting an error if the
user tries to access any data without the trace
being open.