Improve error handling
Normally when a C callback calls back into Python code,
ctypes simply prints and swallows the Python exception.
This results in the same exception being spammed over
and over again, until libundumpi finishes excecution.

To prevent this behaviour we simply wrap every user
callback in an try-except block. If an exception occurs
we manually print the exception trace and abort the
process.