Add time utility methods to DumpiClock
Objects of DumpiClock, which is modelled after C's
'timespec', can now be subtracted and added for easier
computation of time differences.
If time a is subtracted from time b and if a < b, the
resulting time difference will be negative. Since Python
automatically supports arbitrary precision (integer)
arithmetic, there should not be any problems with overflow.

Also added three conversion functions which return a
DumpiClock as a single number in seconds, milliseconds
or nanoseconds, as well as a string representation.